Notification: refactor to more domain-driven directory design
