--- notification/src/interfaces/notificationCallbacks/fcmCallback.ts
@@ -1,8 +1,0 @@
-export default interface FcmCallback {
-  notification: {
-    title: string;
-    body: string;
-  };
-  data?: {[key: string]: string};
-  topic: string;
-}
