--- notification/src/interfaces/notificationCallbacks/slackCallback.ts
@@ -1,3 +1,0 @@
-export default interface SlackCallback {
-  text: string;
-}
