--- notification/src/interfaces/notificationCallbacks/webhookCallback.ts
@@ -1,4 +1,0 @@
-export default interface WebhookCallback {
-  location: string;
-  timestamp: Date;
-}
