+++ notification/src/notification-execution/condition-evaluation/transformationResults/jobResult.ts
